--- docbasesync/__init__.py
@@ -1,1 +1,0 @@
-from docbasesync.app import App  # noqa
