+++ shosai/__init__.py
@@ -1,0 +1,1 @@
+from shosai.app import App  # noqa
