+++ shosai/langhelpers.py
